+++ .cmake-format.py
@@ -1,0 +1,8 @@
+additional_commands = {
+  "pkg_find": {
+    "kwargs": {
+      "PKG": "*"
+    }
+  }
+}
+